TFT35 compatible with both ILI9488 and NT35310
