+++ TFT/src/User/API/UI/ui_draw.c
@@ -14,7 +14,6 @@ void lcd_frame_display(uint16_t sx, uint16_t sy, uint16_t w, uint16_t h, uint32_
   uint32_t address = addr;
 
   LCD_SetWindow(sx, sy, sx + w - 1, sy + h - 1);
-  LCD_WR_REG(TFTLCD_WRITEMEMORY);
 
   W25Qxx_SPI_CS_Set(0);
   W25Qxx_SPI_Read_Write_Byte(CMD_READ_DATA);
@@ -81,7 +80,6 @@ bool model_DirectDisplay(GUI_POINT pos, char *gcode)
   f_lseek(&gcodeFile, gcodeFile.fptr + 3);
 
   LCD_SetWindow(pos.x, pos.y, pos.x + ICON_WIDTH - 1, pos.y + ICON_HEIGHT - 1);
-  LCD_WR_REG(TFTLCD_WRITEMEMORY);
   for (uint16_t y = 0; y < ICON_HEIGHT; y++)
   {
     for (uint16_t x = 0; x < ICON_WIDTH; x++)
@@ -169,7 +167,6 @@ void ICON_PressedDisplay(uint16_t sx, uint16_t sy, uint8_t icon)
   uint32_t address = getBMPsize((uint8_t *)&w, (uint8_t *)&h, ICON_ADDR(icon));
 
   LCD_SetWindow(sx, sy, sx + w - 1, sy + h - 1);
-  LCD_WR_REG(TFTLCD_WRITEMEMORY);
 
   W25Qxx_SPI_CS_Set(0);
   W25Qxx_SPI_Read_Write_Byte(CMD_READ_DATA);
