+++ TFT/src/User/Variants/pin_TFT28_V1_0.h
@@ -20,8 +20,6 @@
 #ifndef TFTLCD_DRIVER
   #define TFTLCD_DRIVER               ILI9341  // Type of LCD driver, now support[RM68042, ILI9488, ILI9341, ST7789, HX8558].
   #define TFTLCD_DRIVER_SPEED         0x03
-  #define TFTLCD_0_DEGREE_REG_VALUE   0X68
-  #define TFTLCD_180_DEGREE_REG_VALUE 0XA8
 #endif
 #ifndef LCD_DATA_16BIT
   #define LCD_DATA_16BIT 1  // LCD data 16bit or 8bit
