+++ TFT/src/User/Variants/pin_TFT28_V3_0.h
@@ -20,8 +20,6 @@
 #ifndef TFTLCD_DRIVER
   #define TFTLCD_DRIVER               ST7789  // Type of LCD driver, now support[RM68042, ILI9488, ILI9341, ST7789, HX8558].
   #define TFTLCD_DRIVER_SPEED         0x05
-  #define TFTLCD_0_DEGREE_REG_VALUE   0X60
-  #define TFTLCD_180_DEGREE_REG_VALUE 0XA0
 #endif
 
 #include "pin_TFT35_E3_V3_0.h"
