+++ TFT/src/User/Variants/pin_TFT35_V1_2.h
@@ -10,8 +10,6 @@
 #ifndef TFTLCD_DRIVER
   #define TFTLCD_DRIVER               ILI9488  // Type of LCD driver, now support[RM68042, ILI9488, ILI9341, ST7789, HX8558].
   #define TFTLCD_DRIVER_SPEED         0x03
-  #define TFTLCD_0_DEGREE_REG_VALUE   0X28
-  #define TFTLCD_180_DEGREE_REG_VALUE 0XE8
 #endif
 #ifndef LCD_DATA_16BIT
   #define LCD_DATA_16BIT 1  // LCD data 16bit or 8bit
