+++ TFT/src/User/Variants/pin_Template.h
@@ -23,8 +23,6 @@
 #ifndef TFTLCD_DRIVER
   #define TFTLCD_DRIVER               RM68042  // Type of LCD driver, now support[RM68042, ILI9488, ILI9341, ST7789, HX8558, SSD1963].
   #define TFTLCD_DRIVER_SPEED         0x03
-  #define TFTLCD_0_DEGREE_REG_VALUE   0X28
-  #define TFTLCD_180_DEGREE_REG_VALUE 0X2B
 #endif
 
 //#ifndef SSD1963_LCD_PARA  // Only for TFTLCD_DRIVER is SSD1963
